pyproject.toml
[tool.poetry]
name = "pysurfex-experiment"
version = "0.1.0"
description = "pysurfex-experiment"
authors = ["Trygve Aspelien <trygveasp@met.no>"]
readme = "README.rst"
packages = [{include = "experiment"}, {include = "experiment_scheduler"}, {include = "experiment_tasks"}, {include = "experiment_setup"}]

[tool.poetry.scripts]
PySurfexExp = "experiment:surfex_exp"
PySurfexExpConfig = "experiment:surfex_exp_config"
PySurfexExpSetup = "experiment_setup:surfex_exp_setup"
SubmitTask = "experiment_scheduler:run_submit_cmd_exp"


[tool.poetry.dependencies]
python = "^3.8"
gridpp = "^0.6.0"
titanlib = "^0.3.3"
datetime = "^4.7"
tomlkit = "^0.11.6"
setuptools = "57.2.0"
demjson = "1.6"
cfunits = "3.2.4"
pysurfex-scheduler = {git = "git@github.com:metno/pysurfex-scheduler.git", branch="feature/deode_similar", develop = true}
pysurfex = {git = "git@github.com:metno/pysurfex.git", develop = true}
troika = {git = "https://git.ecmwf.int/scm/ecsdk/troika.git"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint."FORMAT"]
max-line-length = 120
